Includes nom-mail-parser
Previous version of chouette used a custom made ugly mail parser. This
commit adds the `nom-mail-parser` crate, based on `nom` to parse mail.
It is still not fully functional, but it will be much easier to make a
fully functional mail parser out of this than out of the previous one.